### 2025-01-10 #7646 @editorialbot
[PRE REVIEW]: PyDrugLogics: A Python Package for Predicting Drug Synergies Using Boolean Models
**Submitting author:** @szlaura (<a href="http://orcid.org/0009-0002-4141-0500">Laura Szekeres</a>)
**Repository:** https://github.com/druglogics/pydruglogics
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.8
