### 2025-01-11 #7646 @Kevin-Mattheus-Moerman
- [ ] The following is just a suggestion to improve a sentence in your README, I would rephrase this sentence `PyDrugLogics is a Python package designed for constructing, optimizing Boolean Models and performs in-silico perturbations of the models.` to `PyDrugLogics is a Python package designed for constructing and optimizing Boolean Models, and to perform in-silico perturbations of the models.`
