### 2025-02-26 b11253c
Add CONTRIBUTING and CODE_OF_CONDUCT markdown files, update README.md
Two new files have been added to the project, CONTRIBUTING and CODE_OF_CONDUCT markdown files. The README's first sentence has been improved, Contributing and Code of conduct sections have been added.
