+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,35 @@
+# Code of Conduct - PyDrugLogics
+
+## Our Pledge
+
+We, as contributors and maintainers, are committed to fostering a welcoming, inclusive, and harassment-free community for everyone, regardless of personal characteristics or background.
+
+## Our Standards
+
+### Examples of Positive Behavior
+
+- Demonstrating empathy and kindness toward others.
+- Respecting differing opinions and engaging in constructive dialogue.
+- Taking responsibility for mistakes and striving for improvement.
+- Focusing on the overall well-being of the community.
+
+### Examples of Unacceptable Behavior
+
+- Harassment, abuse, or offensive comments.
+- Publishing private information without explicit permission.
+- Personal attacks, trolling, or inflammatory behavior.
+- Any other actions that are unprofessional or inappropriate.
+
+## Scope
+
+This Code of Conduct applies to all community spaces, including online platforms such as GitHub issues, pull requests, discussions, and communication channels (e.g., Slack, email). It also applies whenever an individual officially represents the project in public settings, both online and offline.
+
+## Enforcement
+
+Instances of inappropriate behavior may be reported to the project maintainers at [szekereslaura98@gmail.com](mailto:szekereslaura98@gmail.com). All reports will be reviewed and investigated promptly and respectfully. The privacy and safety of the reporter will be protected.
+
+Maintainers have the right to remove or reject contributions, comments, or other interactions that do not align with this Code of Conduct. Violations may result in warnings, temporary restrictions, or permanent removal from the community, depending on severity.
+
+## Commitment to Community
+
+Our goal is to build a thriving, inclusive, and collaborative community. We value your participation and thank you for contributing to the success of PyDrugLogics.
