### 2025-02-26 #7646 @szlaura
Hi @Kevin-Mattheus-Moerman,

Thank you for your feedback and patience! I have now updated the project based on your suggestions:

- Added a CONTRIBUTING.md file with guidelines for contributors and linked it in the README.md.

- Updated the sentence in the README.md for better clarity, as per your suggestion.

If you have any further suggestions or feedback, I'd be happy to incorporate them.
