### 2025-04-07 #7646 @abhishektiwari
👋 @clreda @davidegraff @Sulstice @BobAubouin @ArtPoon, would any of you be willing to review this submission for JOSS?  The submission under consideration is "PyDrugLogics: A Python Package for Predicting Drug Synergies Using Boolean Models".

Based on your experience and your past authorship of related JOSS publications, we think you might be able to provide a great review of this submission. If not, could you recommend any potential reviewers?

Thanks
Abhi
