### 2025-04-28 druglogics/pydruglogics#3 @BobAubouin
dependencies in pyproject
Hi @szlaura,

When installing your package, two steps are needed to install the package and its requirements. This could be shortened in one single step by including the requirement in the pyproject.toml file, see [here](https://packaging.python.org/en/latest/guides/writing-pyproject-toml/) for documentation.

It could be also great to set the python version to ensure reproducibility of the results.

Bests,

Bob
