### 2025-04-29 druglogics/pydruglogics#4 @BobAubouin
Improve API documentation
Hi,

Although the tutorial section of the documentation is very complete, the API is a bit less documented and most of the methods lack description. In order to improve the possibility to host contribution, it could be great to improve this part of the documentation.

Also, the formatting seems to be broken in this part of the documentation.

Bests,
Bob
