### 2025-04-29 druglogics/pydruglogics#5 @BobAubouin
Fix test coverage
Hi,

The test coverage seems to be broken as it indicate a 0% of coverage (downloading the last git version and running _pytest --cov=pydruglogics tests_). It could be great to fix it to have access to this indicator.

Bests,
Bob
