### 2025-04-29 #8038 @BobAubouin
Dear authors @szlaura,

I believe that you have made a great job regarding this python package, I particularly acknowledge the great work made on the documentation and the tutorial sections. Globally this submission is adequate, and my comments are only on small improvements. I created issues in your repository for the one related to the code:

- The installation could be simplified by including the requirements of the package in the project file --> druglogics/pydruglogics#3
- The API could be better documented at the method scale. It will be particularly useful for collaboration. -->  druglogics/pydruglogics#4
- The features of the overall package, and whether they are new comparing to the state of the art, could be made more explicit in the paper.
- Although there is a great test pipeline, the coverage test does not seem to work as it indicate a 0% coverage. It could be fixed.  druglogics/pydruglogics#5
- The last one is more a question, why are there two tutorial files, one simple and one for Colomoto, which include the same code ?

Bests,
Bob
