### 2025-04-29 druglogics/pydruglogics#5 @clreda
Dear all,

The coverage indicator is for automated tests run in the workflow—but there are none in run-test.yml, hence the 0%. A way to circumvent this is to execute the files in the folder tests/ in run-test.yml
```bash
cd tests/executor
python3 -m Executor_test
cd ../input
python3 -m ModelOutputs_test
python3 -m Perturbation_test
python3 -m TrainingData_test
cd ../model
python3 -m BooleanModel_test
python3 -m Evolution_test
python3 -m InteractionModel_test
python3 -m ModelPredictions_test
python3 -m Statistics_test
cd ../utils
python3 -m BNetworkUtil_test
python3 -m PlotUtil_test
cd ../../
```

Cheers
