### 2025-04-29 druglogics/pydruglogics#3 @clreda
Dear all,

I would also add the operating system in which the package was implemented and maintained, as it does make a difference in terms of package versioning.

Cheers
