### 2025-04-29 druglogics/pydruglogics#6 @clreda
Error in installation instructions for Conda
Dear all,

When I try to run the installation commands specific to Conda, I encounter an error and the installation is unsuccessful.

**Operating system:** Linux Debian 11
**Python version:** 3.11

- **Minimal example 1:** using Miniconda for managing a virtual environment:

```bash
conda create -n pydruglogics python=3.11 -y
conda activate pydruglogics
conda install szlaura::pydruglogics
```

**Expected output:** Successful installation.
**Observed output:** Error raised:

```bash
Error: PackagesNotFoundError: The following packages are not available from current channels:
  - szlaura::pydruglogics
```

I tried to solve this problem but encountered another error.

- **Minimal example 2:** using Miniconda for managing a virtual environment:

```bash
conda create -n pydruglogics python=3.11 -y
conda activate pydruglogics
conda install -c szlaura pydruglogics ## adding a unseen channel to Conda as the channel by default is defaults
```

**Expected output:** Successful installation.
**Observed output:** Error raised:

```bash
Error: The following package could not be installed
 └─ pydruglogics is not installable because it requires
   └─ mpbn >=3.8 , which does not exist (perhaps a missing channel).
```

Please let me know if you need further details.

Cheers
