### 2025-04-29 druglogics/pydruglogics#7 @clreda
Error in installation instructions with source code
Dear all,

When I try to run the installation commands using the source code, I encounter an error and the installation is unsuccessful.

**Operating system:** Linux Debian 11
**Python version:** 3.11

- **Minimal example:** using Miniconda for managing a virtual environment:

```bash
conda create -n pydruglogics python=3.11 -y
conda activate pydruglogics
git clone https://github.com/druglogics/pydruglogics.git
cd pydruglogics
pip install -r requirements.txt
pip install .
## Running tests to check the installation
cd tests/executor
python3 -m Executor_test
```

**Expected output:** Successful installation and silent completion of the test.
**Observed output:** Error raised:

```bash
Error: ModuleNotFoundError: No module named 'pydruglogics'
```

Please let me know if you need further details.

Cheers
