### 2025-04-29 #8038 @clreda
Dear @szlaura, dear all,

I concur with all that @BobAubouin mentioned, and I added comments/feedback to their issues (automated tests/coverage, API documentation, requirements). My checklist is almost complete so good job! Yet I have also further concerns:
- Installation issues with Conda [#6](https://github.com/druglogics/pydruglogics/issues/6)
- Installation issues with source code [#7](https://github.com/druglogics/pydruglogics/issues/7)
- PyDrugLogics is a Python implementation of DrugLogics, which is a preexisting package implemented in Java. The paper mentions as specific contributions to PyDrugLogics: "[PyDrugLogics] significantly boosts [DrugLogics] by reducing the code complexity, improving the execution time, and introducing new features that expand its capabilities" but the claims about improved execution time and new features are unsubstantiated. What are the new features introduced compared to the Java implementation? Can a small benchmark be performed to compare PyDrugLogics and DrugLogics?

Cheers
