### 2025-06-01 f230af2
Remove requirements.txt and define dependencies in pyproject.toml
Migrated all dependencies to pyproject.toml. Removed requirements.txt to streamline installation. Updated README.md to reflect the new installation process.
