+++ requirements.txt
@@ -1,9 +1,0 @@
-joblib~=1.4.2
-matplotlib~=3.9.2
-mpbn~=3.8
-numpy~=2.1.3
-pandas~=2.2.3
-pygad~=3.3.1
-scikit_learn~=1.5.2
-scipy~=1.14.1
-pyboolnet @ git+https://github.com/hklarner/pyboolnet@3.0.16#egg=pyboolnet
