+++ pyproject.toml
@@ -1,3 +1,35 @@
 [build-system]
-requires = ["setuptools>=42", "wheel"]
+requires = ["setuptools>=62.6", "wheel"]
 build-backend = "setuptools.build_meta"
+
+[project]
+name = "pydruglogics"
+version = "0.1.8"
+description = "A Python package designed for constructing and optimizing Boolean models, and to perform in-silico perturbations of the models."
+readme = "README.md"
+authors = [
+  { name="Laura Szekeres", email="szekereslaura98@gmail.com" }
+]
+requires-python = ">=3.10"
+license = { text = "GPL-3.0-or-later" }
+
+
+dependencies = [
+    "joblib~=1.4.2",
+    "matplotlib~=3.9.2",
+    "mpbn~=3.8",
+    "numpy~=2.1.3",
+    "pandas~=2.2.3",
+    "pygad~=3.3.1",
+    "scikit-learn~=1.5.2",
+    "scipy~=1.14.1",
+    "pyboolnet @ git+https://github.com/hklarner/pyboolnet@3.0.16#egg=pyboolnet"
+]
+
+classifiers = [
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Operating System :: POSIX :: Linux",
+    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
+    "Intended Audience :: Science/Research"
+]
