+++ .github/workflows/run-tests.yml
@@ -25,7 +25,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install pytest pytest-cov
-        pip install -r requirements.txt
         pip install -e .
 
     - name: Run tests
