### 2025-06-01 druglogics/pydruglogics#3 @szlaura
Hi @BobAubouin  and @clreda,

Thank you for your helpful suggestions!

- All dependencies are now listed directly in pyproject.toml, so the package and requirements can be installed in a single step: `pip install .`
- The supported Python version is now set in the pyproject.toml
- Operating system information, and compatibility details are provided via classifiers and mentioned in the README

If you have any further feedback or encounter issues, feel free to open another issue or comment here.

Best,
Laura
