### 2025-06-03 druglogics/pydruglogics#6 @szlaura
Hi [Name],

Thank you so much for your detailed feedback and for taking the time to report your installation experience.

I’ve tested the installation myself in a clean Debian 11 environment with Python 3.11, and in my case, everything worked as expected. Based on your error messages, I believe the problem might be related to Conda not searching all the necessary channels for dependencies like mpbn.

Would you be willing to try the following command, which explicitly specifies all required channels?

`conda install -c szlaura -c colomoto -c conda-forge pydruglogics`

If this still does not work, could you please share the full error output and the result of the following commands?

```
conda config --show channels
conda --version
```

This will help me further investigate any environment-specific issues.

Thank you again for your patience and for helping improve the package!

Best,
Laura
