### 2025-06-04 druglogics/pydruglogics#6 @clreda
Dear @szlaura,

Using the command
``conda install -c szlaura -c colomoto -c conda-forge pydruglogics``
solved the issue.

When I tested the code, I ran into an error in file tests/executor/Executor_test.py
``ModuleNotFoundError: No module named 'pytest``

Installing the missing module solved the issue
``conda install -c conda-forge pytest -y``

Otherwise, all the tests and the example usage code ran successfully. Thank you!

Cheers
