### 2025-06-28 druglogics/pydruglogics#7 @szlaura
Hi @clreda,

Thanks so much for the detailed feedback! Your explanations and environment details were very helpful.

After reviewing your comments, I noticed that the missing `__init__.py` files in the ` tests/ `directories were likely causing the test discovery and import errors.

I've now added `__init__.py ` files to the test directories, that allows to Python to treat them as a proper package. These files sometimes help Python and various development tools recognize test directories as packages, which may address the issues you experienced.

From your example, I see that you were running the test with:

```
cd tests/executor
python3 -m Executor_test
```

With the recent update, I’d like to suggest that you try running the test from the root of the repository instead,  and using:

`python -m tests.executor.Executor_test`

This approach is often more portable and robust, as it leverages the package structure that Python expects.

If you get a chance to try this, I would really appreciate any feedback on whether it works as expected. But if you prefer your original method or still encounter issues, please let me know. I'm here to help troubleshoot further.

Best regards,
Laura
