### 2025-06-28 6e8f587
[fix] Extend pyproject.toml with dependencies and project details, minimize requirements.txt, update README
Moved all essential and test dependencies in pyproject.toml, except pyboolnet, which cannot be included. Added a dedicated [project.optional-dependencies] section. Updated the README according to the new updates.
