+++ pyproject.toml
@@ -1,3 +1,38 @@
 [build-system]
 requires = ["setuptools>=42", "wheel"]
 build-backend = "setuptools.build_meta"
+
+[project]
+name = "pydruglogics"
+version = "0.1.8"
+description = "Constructing, optimizing Boolean Models and performing in-silico perturbations."
+readme = "README.md"
+requires-python = ">=3.10"
+license = { file = "LICENSE" }
+authors = [
+    { name = "Laura Szekeres"}
+]
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
+    "Operating System :: POSIX :: Linux",
+    "Intended Audience :: Science/Research",
+    "Topic :: Scientific/Engineering :: Bio-Informatics",
+]
+
+dependencies = [
+    "joblib~=1.4.2",
+    "matplotlib~=3.9.2",
+    "mpbn~=3.8",
+    "numpy~=2.1.3",
+    "pandas~=2.2.3",
+    "pygad~=3.3.1",
+    "scikit-learn~=1.5.2",
+    "scipy~=1.14.1"
+]
+
+[project.optional-dependencies]
+test = [
+    "pytest",
+    "pytest-cov"
+]
