### 2025-06-28 druglogics/pydruglogics#5 @szlaura
Hi @BobAubouin and @clreda,

Thank for the feedback and suggestions on the test coverage issue.

I’ve now added `pytest` and `pytest-cov` as optional dependencies in `pyproject.toml`, and I’ve updated the [README](https://github.com/druglogics/pydruglogics/blob/main/README.md) with new instructions for installing test dependencies and running the coverage checks.

To get started with testing and coverage, you can now run:

```
    pip install -r requirements.txt
    pip install -e .[test]
```

Then, from the repository root, run:

`pytest tests`

I hope this update helps! If you run into any more problems or have further suggestions regarding to the test coverage, please let me know here.

Best,
Laura
