### 2025-06-28 druglogics/pydruglogics#3 @szlaura
Hi @BobAubouin and @clreda ,

Thank you for the helpful suggestions and for sharing the documentation link.

Here is a little summary of what I updated in the pyproject.toml:

- Added all main package dependencies under the dependencies section
- Introduced a `[project.optional-dependencies]` section for test tools like `pytest` and `pytest-cov`
- Specified the minimum required Python version
- Updated license and authorship information
- Added classifiers for Python version, license, intended audience, and operating system

Unfortunately, since `pyboolnet` is not available on PyPI, I cannot include it in `pyproject.toml `. Instead, it remains in requirements.txt, as `pip` can install it directly from its GitHub repository using that file.

If anyone has a solution for including GitHub-based dependencies directly in `pyproject.toml`, I’d be very interested to hear about it. Please let me know if you have any other suggestions or feedback.

Thank you again for your feedback and support.

Best regards,
Laura
