### 2025-06-28 #8038 @szlaura
Hi all,

I wanted to share a quick update on the recent progress.
I have addressed several of the open issues:
- The installation that could be simplified: https://github.com/druglogics/pydruglogics/issues/3
- The coverage test issue: https://github.com/druglogics/pydruglogics/issues/5
- Installation issues with source code:  https://github.com/druglogics/pydruglogics/issues/7

There is still one issue related to [API documentation](https://github.com/druglogics/pydruglogics/issues/4) that I am actively working on. In addition, I am making ongoing improvements to the paper.

Thank you all for your feedback and suggestions, they have been very helpful.

Best,
Laura
