### 2025-06-30 druglogics/pydruglogics#7 @clreda
Dear @szlaura,

With the newest commit, both versions now run:

```bash
conda create -n pydruglogics python=3.11 -y
conda activate pydruglogics
git clone https://github.com/druglogics/pydruglogics.git
cd pydruglogics
pip install -r requirements.txt
pip install .
## Running tests to check the installation : approach 1
python -m tests.executor.Executor_test
## Running tests to check the installation : approach 2
cd tests/executor
python3 -m Executor_test
```

Note that with Approach 1 only, the following warning appears:
```bash
INFO Failed to extract font properties from /usr/share/fonts/opentype/cantarell/Cantarell-VF.otf: In FT2Font: Can not load face (SFNT font table missing; error code 0x8e)
INFO Failed to extract font properties from /usr/share/fonts/truetype/noto/NotoColorEmoji.ttf: In FT2Font: Can not load face (unknown file format; error code 0x2)
INFO generated new fontManager
```
but this does not prevent the code from running.

Cheers,
