### 2025-07-17 149ed36
[fix] Extend pyproject.toml with dependency, update README
Moved pyboolnet dependency to pyproject.toml and updated the new install instructions in the README.
