+++ requirements.txt
@@ -1,1 +1,0 @@
-pyboolnet @ git+https://github.com/hklarner/pyboolnet@3.0.16#egg=pyboolnet
