### 2025-07-18 7da3d09
Update APi documentation, install instruction
Added a more detailed description for the API, updated the installation instructions and the version.
