+++ docs/conf.py
@@ -11,7 +11,7 @@ import sys
 project = 'PyDrugLogics'
 copyright = '2024, Laura Szekeres'
 author = 'Laura Szekeres'
-version = '0.1.8'
+version = '0.1.9'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
@@ -30,6 +30,7 @@ autoclass_content = 'both'
 templates_path = ['_templates']
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 source_suffix = ['.rst', '.md']
+autodoc_typehints = "none"
 
 sys.path.insert(0, os.path.abspath('../'))
 
