+++ docs/pydruglogics.input.ModelOutputs.rst
@@ -1,7 +1,19 @@
-pydruglogics.input.ModelOutputs module
-======================================
+pydruglogics.input.ModelOutputs
+===============================
 
-.. automodule:: pydruglogics.input.ModelOutputs
-   :members:
-   :undoc-members:
-   :show-inheritance:
+Handles model outputs for each node, including loading from file or dictionary,
+computing min/max output, and exposing accessors.
+
+.. automethod:: pydruglogics.input.ModelOutputs.ModelOutputs.get_model_output
+
+    Get the output weight for a given node.
+
+.. automethod:: pydruglogics.input.ModelOutputs.ModelOutputs.size
+
+    Return the number of outputs (nodes).
+
+
+.. automethod:: pydruglogics.input.ModelOutputs.ModelOutputs.print
+
+    Print the string representation of the ModelOutputs instance.
+
