+++ docs/pydruglogics.model.BooleanModelOptimizer.rst
@@ -1,7 +1,15 @@
-pydruglogics.model.BooleanModelOptimizer module
-===============================================
+pydruglogics.model.BooleanModelOptimizer
+========================================
 
-.. automodule:: pydruglogics.model.BooleanModelOptimizer
-   :members:
-   :undoc-members:
-   :show-inheritance:
+Abstract base class for optimizers that train Boolean models and can save the results.
+
+.. automethod:: pydruglogics.model.BooleanModelOptimizer.BooleanModelOptimizer.run
+    :noindex:
+
+    Abstract method. Run the optimizer and return a list of trained BooleanModel instances.
+
+
+.. automethod:: pydruglogics.model.BooleanModelOptimizer.BooleanModelOptimizer.save_to_file_models
+
+    Abstract method. Save the optimized models to the given path.
+
