+++ docs/pydruglogics.model.ModelPredictions.rst
@@ -1,7 +1,17 @@
-pydruglogics.model.ModelPredictions module
-==========================================
+pydruglogics.model.ModelPredictions
+===================================
 
-.. automodule:: pydruglogics.model.ModelPredictions
-   :members:
-   :undoc-members:
-   :show-inheritance:
+Simulates Boolean models under perturbations, computes responses, and calculates synergy scores using HSA or Bliss methods.
+
+
+.. automethod:: pydruglogics.model.ModelPredictions.ModelPredictions.run_simulations
+
+
+.. automethod:: pydruglogics.model.ModelPredictions.ModelPredictions.save_to_file_predictions
+
+    Save the prediction matrix and synergy results to disk.
+
+
+.. automethod:: pydruglogics.model.ModelPredictions.ModelPredictions.get_prediction_matrix
+
+    Print or retrieve the prediction matrix (perturbed responses).

[…]
