### 2025-07-18 a8424ec
New package version 0.1.9.
Updated version from 0.1.8 to 0.1.9 in the project.
