+++ pydruglogics/__init__.py
@@ -16,7 +16,7 @@ logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 
-__version__ = "0.1.8"
+__version__ = "0.1.9"
 
 __all__ = [
     "Perturbation",
