+++ conda/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "pydruglogics" %}
-{% set version = environ.get('VERSION', '0.1.8') %}
+{% set version = environ.get('VERSION', '0.1.9') %}
 
 package:
   name: {{ name|lower }}
