### 2025-07-18 druglogics/pydruglogics#4 @BobAubouin
Hi Laura,

It looks really great!

I just noticed that there are still some formatting issues with some of the page (such as statistic for instance).

Best,
Bob
