### 2025-07-18 0eb1728
Update API documentation with more structured descriptions
Added a more detailed and organized description for the API.
