+++ docs/pydruglogics.model.BooleanModel.rst
@@ -5,18 +5,58 @@ A Boolean network model with support for mutation, attractor calculation, pertur
 
 .. automethod:: pydruglogics.model.BooleanModel.BooleanModel.calculate_attractors
 
+   Calculate attractors based on the chosen attractor_tool and attractor_type.
+
+   **Parameters**
+
+   - ``attractor_tool (str)``: Tool for attractor calculation. Possible values: ``'mpbn'``, ``'pyboolnet'``.
+   - ``attractor_type (str)``: Type of attractor. Possible values: ``'stable_states'``, ``'trapspaces'``.
 
 .. automethod:: pydruglogics.model.BooleanModel.BooleanModel.calculate_global_output
 
+   Calculates the (normalized) global output of the model.
+
+   **Parameters**
+
+   - ``model_outputs (ModelOutputs)``: Node weights.
+   - ``normalized (bool, optional)``: Whether to normalize. Default: True.
+
+   **Returns**
+
+   - float: The (normalized) global output.
 
 .. automethod:: pydruglogics.model.BooleanModel.BooleanModel.from_binary
 
+   Updates the Boolean Equations from a binary representation.
+
+   **Parameters**
+
+   - ``binary_representation (list)``: The binary representation of the Boolean Equations.
+   - ``mutation_type (str)``: Mutation type: ``'topology'``, ``'balanced'``, or ``'mixed'``.
+
+   **Returns**
+
+   - list: Updated Boolean Equations.
 
 .. automethod:: pydruglogics.model.BooleanModel.BooleanModel.to_binary
 
+   Converts the Boolean Equations to a binary representation.
+
+   **Parameters**
+
+   - ``mutation_type (str)``: The type of mutation.
+
+   **Returns**
+
+   - list: Binary representation of Boolean Equations.
 
 .. automethod:: pydruglogics.model.BooleanModel.BooleanModel.add_perturbations
 
+   Adds perturbations to the Boolean Model.
+
+   **Parameters**
+
+   - ``perturbations (list)``: List of perturbations.
 
 .. automethod:: pydruglogics.model.BooleanModel.BooleanModel.print
 
