### 2025-07-28 druglogics/pydruglogics#3 @clreda
Dear @szlaura,

I agree with your approach.

Cheers,
