### 2025-08-14 ac52fbd
[fix] Restore requirements.txt file and add to the project
Due to the PyPi external package restrictions, the PyBoolNet package is installed from requirements.txt file. Updated the README.md and pyproject.toml.
