### 2025-08-14 1e875d6
Update the Getting Started page with the updated installation process
Added the requirements.txt to the installation steps.
