### 2025-08-15 dd03bf5
Extend JOSS paper with new section (pipeline evaluation)
Added detailed comparison of PyDrugLogics and DrugLogics, illustrating performance with figures and examples.
