+++ paper/paper.bib
@@ -30,6 +30,22 @@
   keywords = {Animals,Drug Synergism,Humans,Models Biological}
 }
 
+@article{blissTOXICITYPOISONSAPPLIED1939,
+  title = {{{THE TOXICITY OF POISONS APPLIED JOINTLY1}}},
+  author = {Bliss, C. I.},
+  year = {1939},
+  month = aug,
+  journal = {Annals of Applied Biology},
+  volume = {26},
+  number = {3},
+  pages = {585--615},
+  issn = {00034746},
+  doi = {10.1111/j.1744-7348.1939.tb06990.x},
+  urldate = {2025-08-15},
+  copyright = {http://doi.wiley.com/10.1002/tdm\_license\_1.1},
+  langid = {english}
+}
+
 @misc{chatainMostPermissiveSemantics2018,
   title = {Most {{Permissive Semantics}} of {{Boolean Networks}}},
   author = {Chatain, Thomas and Haar, Stefan and Kol\{{\v c}\}{\'a}k, Juraj and Paulev{\'e}, Lo{\"i}c},
@@ -71,6 +87,23 @@
   doi = {10.15252/msb.20188664},
   urldate = {2024-12-13},
   langid = {english},
+}
+
+@article{flobakDiscoveryDrugSynergies2015,
+  title = {Discovery of {{Drug Synergies}} in {{Gastric Cancer Cells Predicted}} by {{Logical Modeling}}},
+  author = {Flobak, {\AA}smund and Baudot, Ana{\"i}s and Remy, Elisabeth and Thommesen, Liv and Thieffry, Denis and Kuiper, Martin and L{\ae}greid, Astrid},
+  editor = {Xenarios, Ioannis},
+  year = {2015},
+  month = aug,
+  journal = {PLOS Computational Biology},
+  volume = {11},
+  number = {8},
+  pages = {e1004426},
+  issn = {1553-7358},
+  doi = {10.1371/journal.pcbi.1004426},
+  urldate = {2025-08-15},
+  langid = {english},
+  file = {/home/laura/Zotero/storage/U9VYCI7Y/Flobak et al. - 2015 - Discovery of Drug Synergies in Gastric Cancer Cells Predicted by Logical Modeling.pdf}
 }
 
 @article{flobakFineTuningLogical2023,
@@ -194,6 +227,16 @@
   urldate = {2024-12-04},
   abstract = {Abstract                            Background               Cells process signals using complex and dynamic networks. Studying how this is performed in a context and cell type specific way is essential to understand signaling both in physiological and diseased situations. Context-specific medium/high throughput proteomic data measured upon perturbation is now relatively easy to obtain but formalisms that can take advantage of these features to build models of signaling are still comparatively scarce.                                         Results                                Here we present                 CellNOptR                 , an open-source R software package for building predictive logic models of signaling networks by training networks derived from prior knowledge to signaling (typically phosphoproteomic) data.                 CellNOptR                 features different logic formalisms, from Boolean models to differential equations, in a common framework. These different logic model representations accommodate state and time values with increasing levels of detail. We provide in addition an interface via Cytoscape (                 CytoCopteR                 ) to facilitate use and integration with Cytoscape network-based capabilities.                                                        Conclusions                                Models generated with this pipeline have two key features. First, they are constrained by prior knowledge about the network but trained to data. They are therefore context and cell line specific, which results in enhanced predictive and mechanistic insights. Second, they can be built using different logic formalisms depending on the richness of the available data. Models built with                 CellNOptR                 are useful tools to understand how signals are processed by cells and how this is altered in disease. They can be used to predict the effect of perturbations (individual or in combinations), and pote
[…]
