### 2025-08-16 7069829
Move detailed pipeline evaluation to wiki page and update the JOSS paper
Relocated performance comparison between PyDrugLogics and DrugLogics to project wiki. Updated paper's Statement of Need section to include reference to the new wiki section.
