+++ paper/paper.bib
@@ -271,3 +271,10 @@ year = {2023}
   note = {PyPi package version 0.1.8},
   url = {https://pypi.org/project/pydruglogics/}
 }
+
+@misc{pydl-wiki,
+  author = {Szekeres, Laura},
+  title = {Pipeline Evaluation of PyDrugLogics},
+  year = {2025},
+  url = {https://github.com/druglogics/pydruglogics/wiki/Pipeline-Evaluation-of-PyDrugLogics}
+}
