+++ paper/paper.md
@@ -72,7 +72,8 @@ and introducing new features that expand its capabilities. By unifying the funct
 a single Python package, PyDrugLogics simplifies the installation and software maintenance. Example improvements include 
 the use of a standardized format, BoolNet [@musselBoolNetPackageGeneration2010], for loading the models, as well as 
 visualization options (i.e. precision-recall (PR) and receiver operating characteristic (ROC) curves) and statistical 
-analyses (i.e. repeated subsampling of the ensemble Boolean models) for robust evaluation of prediction performance. 
+analyses (i.e. repeated subsampling of the ensemble Boolean models) for robust evaluation of prediction performance.
+Additional examples and comparisons of the Python and Java pipeline are available on the project wiki [@pydl-wiki].
 PyDrugLogics provides an easy-to-use, flexible, and up-to-date solution for simulating Boolean networks and predicting 
 synergistic drug combinations for the prioritization of follow-up lab experiments.
 
@@ -104,48 +105,6 @@ synergies, which serve as the ground truth, and they are typically derived from 
 sources. Using the observed synergies (binary labels: 0 for non-synergistic and 1 for synergistic) as the ground truth, 
 binary classification metrics such as the ROC and PR AUC (area under the curve) are generated to evaluate how well 
 the predicted synergy scores distinguish between synergistic and non-synergistic drug combinations.
-
-# Pipeline Evaluation
-
-The performance and reliability of PyDrugLogics and DrugLogics were compared using the CASCADE 1.0 signaling network 
-and the experimentally observed synergy labels from @flobakDiscoveryDrugSynergies2015.
-
-The execution times were measured using Hyperfine [@Peter_hyperfine_2023] across 10 runs. Both pipelines calculated the 
-trap spaces of the CASCADE 1.0 network while applying topology mutations on it in the genetic algorithm, and synergy 
-scores were calculated using the Bliss method [@blissTOXICITYPOISONSAPPLIED1939]. The evolutionary algorithm performed 
-50 simulations of 20 generations each, using 4 CPU cores. 
-
-The runtime comparison, summarized in Table 1, indicates that PyDrugLogics completes the pipeline more efficiently than 
-DrugLogics. The mean execution time over 10 runs for DrugLogics was 89.030 s ± 1.578 s, whereas PyDrugLogics achieved a 
-mean runtime of 33.239 s ± 3.082 s, demonstrating a clear speed advantage.
-
-| Metric              | DrugLogics (Java)         | PyDrugLogics (Python)     |
-|---------------------|--------------------------|--------------------------|
-| Runs                | 10                        | 10                        |
-| Time (mean ± sigma) | 89.030 s ± 1.578 s      | 33.239 s ± 3.082 s  |
-| Range (min … max)   | 86.051 s … 91.164 s      | 30.577 s … 35.901 s           |
-
-Table: Runtime comparison of DrugLogics and PyDrugLogics across 10 runs.
-
-The ROC and PR curves were generated from the calculated synergy scores against the experimental labels, and the 
-corresponding AUC values were used to evaluate the model performance. Figure 1 shows the performance of DrugLogics and 
-Figure 2 presents PyDrugLogics. Overall, the Python implementation shows slightly better consistency with the 
-experimental data.
-
-![ROC and PR AUC curves of calculated synergy scores for DrugLogics (Java)](figures/Figure1_DrugLogics_ROC_PR.png)
-
-
-![ROC and PR AUC curves of calculated synergy scores for PyDrugLogics (Python)](figures/Figure2_PyDrugLogics_ROC_PR.png)
-
-
-To further quantify the Python pipeline’s robustness, PyDrugLogics performs ensemble subsampling of the Boolean models. 
[…]
