### 2025-08-16 #8038 @szlaura
Hi @abhishektiwari ,

Thanks for the suggestion! I’ve updated the paper accordingly and moved the detailed performance comparison to a dedicated wiki page. Let me know if any further adjustments are needed.

Best,
Laura
