### 2025-08-16 3e2bdaa
Fix missing DOI in JOSS paper references
Added DOI for the reference “What Is Synergy?” by Berenbaum (1989) in paper.bib to resolve JOSS DOI check warning.
