+++ paper/paper.bib
@@ -25,6 +25,7 @@
   number = {2},
   pages = {93--141},
   issn = {0031-6997},
+  doi = {10.1016/S0031-6997(25)00026-2},
   langid = {english},
   pmid = {2692037},
   keywords = {Animals,Drug Synergism,Humans,Models Biological}
