### 2025-08-16 fc6296e
Fix typos in the JOSS paper
Updated the text, corrected spelling and typo errors in the JOSS paper.
