+++ paper/paper.bib
@@ -49,7 +49,7 @@
 
 @misc{chatainMostPermissiveSemantics2018,
   title = {Most {{Permissive Semantics}} of {{Boolean Networks}}},
-  author = {Chatain, Thomas and Haar, Stefan and Kol\{{\v c}\}{\'a}k, Juraj and Paulev{\'e}, Lo{\"i}c},
+  author = {Chatain, Thomas and Haar, Stefan and Kol{\v{c}}{\'a}k, Juraj and Paulev{\'e}, Lo{\"i}c},
   year = {2018},
   publisher = {arXiv},
   doi = {10.48550/ARXIV.1808.10240},
