+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup, find_packages
 
 setup(
     name="pydruglogics",
-    version="0.1.8",
+    version="0.1.9",
     author="Laura Szekeres",
     author_email="szekereslaura98@gmail.com",
     description="PyDrugLogics: a Python package designed for constructing, optimizing Boolean models and performs in-silico perturbations of the models.",
