+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pydruglogics"
-version = "0.1.8"
+version = "0.1.9"
 description = "Constructing, optimizing Boolean Models and performing in-silico perturbations."
 readme = "README.md"
 requires-python = ">=3.10"
