### 2025-08-21 #8038 @abhishektiwari
@szlaura Can you make sure Zenodo author information:

- Formatted as Last Name, First Name
- Includes affiliation and ORCID.

See [an example here](https://zenodo.org/records/16759414)

You need to edit existing version.
