### 2025-08-21 #8038 @abhishektiwari
@szlaura Everything looks good except the `Summary` section of paper which is bit long but more importantly lacks a flow to describe the high-level functionality and purpose of the software for a diverse, non-specialist audience.

I will start with problem (current paragraph 2), then describe how PyDrugLogic solves the problem (paragraph 1).

> For complex diseases such as cancer, combined drug therapies can enhance the efficacy of the personalized treatment and minimize side effects [@jinRationalCombinationsTargeted2023]. Combined drug therapies may have a stronger effect than single drug treatments, a concept referred to as synergy [@berenbaumWhatSynergy1989]. However, finding synergistic drug combinations is a challenging area of modern medicine. The vast number of possible drug combinations, even for small sets of drugs, makes it a complex problem due to the exponential growth in combinations. This fact causes longer testing time in the laboratory experiments and a significant amount of experimental costs.
>
> PyDrugLogics is a Python package that generates optimized Boolean models that represent a biological system and performs in-silico perturbations of these models to predict synergistic drug combinations.

Maybe paragraph 3 and 4 from `Summary` can be moved in `Statement of Need`. Bare minimum we can just rearrange paragraph 1 and 2. See a few example summaries from recently accepted papers [here](https://joss.theoj.org/papers/10.21105/joss.08504), [here](https://joss.theoj.org/papers/10.21105/joss.08076), and [here](https://joss.theoj.org/papers/10.21105/joss.08197).

Can you please review and see if you can improve the `Summary`? Let me know your thoughts.
