### 2025-08-25 592229a
Improve the Summary and Statement of Need section in the JOSS paper.
The Summary section has been shortened, and the Statement of Need section has been extended with a description about logical modeling.
