### 2025-08-25 #8038 @szlaura
Hi @abhishektiwari,

Thank you so much for the great suggestions! I’ve rearranged the paragraphs, shortened the Summary, and moved the detailed modeling explanation to the Statement of Need. Let me know if you notice any more possible improvements.

Best,
Laura
