### 2025-09-17 b9b82f9
Update paper.md, paper.bib with minor typo corrections, add contributing section to README.md
Corrected typos and applied wording changes on the JOSS paper. Added contributing guidelines section to the README.
