### 2025-09-22 354a692
Correct typos in the JOSS paper
Small typo in the Statement of Need and in the References have been corrected.
