+++ paper/paper.bib
@@ -48,7 +48,7 @@
 }
 
 @misc{chatainMostPermissiveSemantics2018,
-  title = {Most {{Permissive Semantics}} of {{Boolean Networks}}},
+  title = {Most Permissive Semantics of {{Boolean}} Networks},
   author = {Chatain, Thomas and Haar, Stefan and Kol{\v{c}}{\'a}k, Juraj and Paulev{\'e}, Lo{\"i}c},
   year = {2018},
   publisher = {arXiv},
@@ -91,7 +91,7 @@
 }
 
 @article{flobakDiscoveryDrugSynergies2015,
-  title = {Discovery of {{Drug Synergies}} in {{Gastric Cancer Cells Predicted}} by {{Logical Modeling}}},
+  title = {Discovery of Drug Synergies in Gastric Cancer Cells Predicted by Logical Modeling},
   author = {Flobak, {\AA}smund and Baudot, Ana{\"i}s and Remy, Elisabeth and Thommesen, Liv and Thieffry, Denis and Kuiper, Martin and L{\ae}greid, Astrid},
   editor = {Xenarios, Ioannis},
   year = {2015},
@@ -186,7 +186,7 @@
 }
 
 @article{naldiCoLoMoToInteractiveNotebook2018,
-  title = {The {{CoLoMoTo Interactive Notebook}}: {{Accessible}} and {{Reproducible Computational Analyses}} for {{Qualitative Biological Networks}}},
+  title = {The {{CoLoMoTo}} Interactive Notebook: {{Accessible}} and Reproducible Computational Analyses for Qualitative Biological Networks},
   shorttitle = {The {{CoLoMoTo Interactive Notebook}}},
   author = {Naldi, Aur{\'e}lien and Hernandez, C{\'e}line and Levy, Nicolas and Stoll, Gautier and Monteiro, Pedro T. and Chaouiya, Claudine and Helikar, Tom{\'a}{\v s} and Zinovyev, Andrei and Calzone, Laurence and {Cohen-Boulakia}, Sarah and Thieffry, Denis and Paulev{\'e}, Lo{\"i}c},
   year = {2018},
@@ -200,7 +200,7 @@
 }
 
 @article{niederdorferStrategiesEnhanceLogic2020,
-  title = {Strategies to {{Enhance Logic Modeling-Based Cell Line-Specific Drug Synergy Prediction}}},
+  title = {Strategies to Enhance Logic Modeling-Based Cell Line-Specific Drug Synergy Prediction},
   author = {Niederdorfer, Barbara and Tour{\'e}, Vasundra and Vazquez, Miguel and Thommesen, Liv and Kuiper, Martin and L{\ae}greid, Astrid and Flobak, {\AA}smund},
   year = {2020},
   month = jul,
@@ -265,7 +265,7 @@ year = {2023}
 
 @misc{pydl-pypi,
   author = {Szekeres, Laura},
-  title = {PyDrugLogics {PyPI} package website},
+  title = {PyDrugLogics {{PyPI}} package website},
   year = {2024},
   publisher = {PyPI},
   note = {PyPI package version 0.1.8},
@@ -274,7 +274,7 @@ year = {2023}
 
 @misc{pydl-wiki,
   author = {Szekeres, Laura},
-  title = {Pipeline Evaluation of PyDrugLogics},
+  title = {Pipeline Evaluation of {{PyDrugLogics}}},
   year = {2025},
   url = {https://github.com/druglogics/pydruglogics/wiki/Pipeline-Evaluation-of-PyDrugLogics}
 }
