### 2025-10-04 #8038 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
