### 2025-10-20 #8038 @rkurchin
Hi @szlaura, I think you should be able to just update the paper (and the Zenodo archive) accordingly and then I can reaccept. Go ahead and make those updates, I'll do that on my end, and then give you the chance to check that everything looks AOK.
