### 2025-10-21 #8038 @szlaura
Hi @rkurchin,
I’ve updated the paper and the Zenodo archive to reflect the complete and correct authorship of the work.
The  Zenodo DOI is [10.5281/zenodo.17410285](https://doi.org/10.5281/zenodo.17410285).

Thank you again for your understanding and support!
